Fixed the quickSort method to not sort the array by reference
The public method quickSort was exposed to work by reference which is
directly against the requirements.  Java treats arrays referencially so
a copy of the array 'sorted' is created to be returned as a seperate
array.  The method would now return the sorted array while leaving the
original array unchanged.