Removed a redundancy of the sort algorithm and re-organized, started assignment 2
Removed the redundant check for j == leftIndex to break the right to
left sweep loop on line 27.  If the left to right sweep loop ever
reaches the left index it has reached our leftmost (checking value)
which will never be less than the value of the index were checking for
the sweep.  Also rearrange the packages to break things out between the
different assignments different pieces of code.  Started getting things
out for assignment 2.  Figured this situation it makes sense to have a
parent class and then have Node be a subclass of that since Node type
will only be used by the AkirixSearchTree anyways.